pyproject.toml
[project]
name = "agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.56.0",
    "livekit-agents[silero,turn-detector]~=1.2",
    "livekit-plugins-google~=0.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
]
